tests/wpt/mozilla/tests/bluetooth/getCharacteristic/disconnect-called-before.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => {
        return gattServer.getPrimaryService(generic_access.name)
        .then(service => {
            gattServer.disconnect();
            return promise_rejects_dom(t, 'NetworkError', service.getCharacteristic(device_name.uuid));
        });
    });
}, 'disconnect() called before getCharacteristic. Reject with NetworkError.');
</script>
